pyproject.toml
[tool.poetry]
name = "pageindexer"
version = "0.1.0"
description = "PageIndexer reads all Shopify products from merchants and index them in a ElasticSearch instance."
authors = ["Lilian Schall <lilian.schall@epita.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
elasticsearch = "^8.12.0"
ShopifyAPI = "^12.4.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
